requirements.txt
Flask==3.0.3
spacy==3.8.5
pandas==2.2.3
playwright==1.51.0
werkzeug==3.0.3
aiohttp==3.11.16
requests==2.31.0
lxml==5.3.2
python-dotenv==1.0.0
flask-cors==4.0.0
flask-restful==0.3.10
PyJWT==2.8.0